sql: move span bounds into helper
This section of query bound generation is reusable and can be refactored
into a a helper function.

Epic: CRDB-58692
Part of: #154551
Part of: #155472
Part of: #161279
Part of: #161276

Release note: None
